refactor: upgrade，deploy

+++ packages/libs/types/entrance.ts
@@ -48,5 +48,5 @@ export function isTransactionVerifier(): i32 {
 }
 
 export function allocate(size: i32): i32 {
-    return heap.alloc(size);
+    return heap.alloc(size) as i32;
 }
